+++ src/Symfony/Component/Intl/Intl.php
@@ -113,14 +113,18 @@ final class Intl
      * Returns the bundle containing currency information.
      *
      * @return CurrencyBundleInterface The currency resource bundle
+     *
+     * @deprecated since Symfony 4.3, to be removed in 5.0. Use {@see Currencies} instead.
      */
     public static function getCurrencyBundle()
     {
+        @trigger_error(sprintf('The method "%s()" is deprecated since Symfony 4.3, use "%s" instead.', __METHOD__, Currencies::class), E_USER_DEPRECATED);
+
         if (null === self::$currencyBundle) {
             self::$currencyBundle = new CurrencyBundle(
                 self::getDataDirectory().'/'.self::CURRENCY_DIR,
                 self::getEntryReader(),
-                self::getLocaleBundle()
+                self::$localeBundle ?? self::$localeBundle = new LocaleBundle(self::getDataDirectory().'/'.self::LOCALE_DIR, self::getEntryReader())
             );
         }
 
@@ -131,14 +135,18 @@ final class Intl
      * Returns the bundle containing language information.
      *
      * @return LanguageBundleInterface The language resource bundle
+     *
+     * @deprecated since Symfony 4.3, to be removed in 5.0. Use {@see Languages} or {@see Scripts} instead.
      */
     public static function getLanguageBundle()
     {
+        @trigger_error(sprintf('The method "%s()" is deprecated since Symfony 4.3, use "%s" or "%s" instead.', __METHOD__, Languages::class, Scripts::class), E_USER_DEPRECATED);
+
         if (null === self::$languageBundle) {
             self::$languageBundle = new LanguageBundle(
                 self::getDataDirectory().'/'.self::LANGUAGE_DIR,
                 self::getEntryReader(),
-                self::getLocaleBundle(),
+                self::$localeBundle ?? self::$localeBundle = new LocaleBundle(self::getDataDirectory().'/'.self::LOCALE_DIR, self::getEntryReader()),
                 new ScriptDataProvider(
                     self::getDataDirectory().'/'.self::SCRIPT_DIR,
                     self::getEntryReader()
@@ -153,9 +161,13 @@ final class Intl
      * Returns the bundle containing locale information.
      *
      * @return LocaleBundleInterface The locale resource bundle
+     *
+     * @deprecated since Symfony 4.3, to be removed in 5.0. Use {@see Locales} instead.
      */
     public static function getLocaleBundle()
     {
+        @trigger_error(sprintf('The method "%s()" is deprecated since Symfony 4.3, use "%s" instead.', __METHOD__, Locales::class), E_USER_DEPRECATED);
+
         if (null === self::$localeBundle) {
             self::$localeBundle = new LocaleBundle(
                 self::getDataDirectory().'/'.self::LOCALE_DIR,
@@ -170,14 +182,18 @@ final class Intl
      * Returns the bundle containing region information.
      *
      * @return RegionBundleInterface The region resource bundle
+     *
+     * @deprecated since Symfony 4.3, to be removed in 5.0. Use {@see Regions} instead.
      */
     public static function getRegionBundle()
     {
+        @trigger_error(sprintf('The method "%s()" is deprecated since Symfony 4.3, use "%s" instead.', __METHOD__, Regions::class), E_USER_DEPRECATED);
+
         if (null === self::$regionBundle) {
             self::$regionBundle = new RegionBundle(
                 self::getDataDirectory().'/'.self::REGION_DIR,
                 self::getEntryReader(),
-                self::getLocaleBundle()
+                self::$localeBundle ?? self::$localeBundle = new LocaleBundle(self::getDataDirectory().'/'.self::LOCALE_DIR, self::getEntryReader())
             );
         }