+++ src/Symfony/Component/Intl/ResourceBundle/ResourceBundleInterface.php
@@ -15,6 +15,8 @@ namespace Symfony\Component\Intl\ResourceBundle;
  * Gives access to ICU data.
  *
  * @author Bernhard Schussek <bschussek@gmail.com>
+ *
+ * @deprecated since Symfony 4.3, to be removed in 5.0.
  */
 interface ResourceBundleInterface
 {