+++ src/Symfony/Component/Intl/Tests/ScriptsTest.php
@@ -9,18 +9,15 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\Component\Intl\Tests\Data\Provider;
+namespace Symfony\Component\Intl\Tests;
 
-use Symfony\Component\Intl\Data\Provider\ScriptDataProvider;
-use Symfony\Component\Intl\Intl;
 use Symfony\Component\Intl\Locale;
+use Symfony\Component\Intl\Scripts;
 
 /**
- * @author Bernhard Schussek <bschussek@gmail.com>
  * @group intl-data
- * @group legacy
  */
-abstract class AbstractScriptDataProviderTest extends AbstractDataProviderTest
+class ScriptsTest extends ResourceBundleTestCase
 {
     // The below arrays document the state of the ICU data bundled with this package.
 
@@ -217,26 +214,9 @@ abstract class AbstractScriptDataProviderTest extends AbstractDataProviderTest
         'Zzzz',
     ];
 
-    /**
-     * @var ScriptDataProvider
-     */
-    protected $dataProvider;
-
-    protected function setUp()
-    {
-        parent::setUp();
-
-        $this->dataProvider = new ScriptDataProvider(
-            $this->getDataDirectory().'/'.Intl::SCRIPT_DIR,
-            $this->createEntryReader()
-        );
-    }
-
-    abstract protected function getDataDirectory();
-
     public function testGetScripts()
     {
-        $this->assertSame(static::$scripts, $this->dataProvider->getScripts());
+        $this->assertSame(self::$scripts, Scripts::getScriptCodes());
     }
 
     /**
@@ -244,7 +224,7 @@ abstract class AbstractScriptDataProviderTest extends AbstractDataProviderTest
      */
     public function testGetNames($displayLocale)
     {
-        $scripts = array_keys($this->dataProvider->getNames($displayLocale));
+        $scripts = array_keys(Scripts::getNames($displayLocale));
 
         sort($scripts);
 
@@ -258,10 +238,7 @@ abstract class AbstractScriptDataProviderTest extends AbstractDataProviderTest
     {
         Locale::setDefault('de_AT');
 
-        $this->assertSame(
-            $this->dataProvider->getNames('de_AT'),
-            $this->dataProvider->getNames()
-        );
+        $this->assertSame(Scripts::getNames('de_AT'), Scripts::getNames());
     }
 
     /**
@@ -272,10 +249,7 @@ abstract class AbstractScriptDataProviderTest extends AbstractDataProviderTest
         // Can't use assertSame(), because some aliases contain scripts with
         // different collation (=order of output) than their aliased locale
         // e.g. sr_Latn_ME => sr_ME
-        $this->assertEquals(
-            $this->dataProvider->getNames($ofLocale),
-            $this->dataProvider->getNames($alias)
-        );
+        $this->assertEquals(Scripts::getNames($ofLocale), Scripts::getNames($alias));
     }
 
     /**
@@ -283,10 +257,10 @@ abstract class AbstractScriptDataProviderTest extends AbstractDataProviderTest
      */
     public function testGetName($displayLocale)
     {
-        $names = $this->dataProvider->getNames($displayLocale);
+        $names = Scripts::getNames($displayLocale);
 
         foreach ($names as $script => $name) {
-            $this->assertSame($name, $this->dataProvider->getName($script, $displayLocale));
+            $this->assertSame($name, Scripts::getName($script, $displayLocale));
         }
     }
 
@@ -294,10 +268,10 @@ abstract class AbstractScriptDataProviderTest extends AbstractDataProviderTest
     {
         Locale::setDefault('de_AT');
 
-        $names = $this->dataProvider->getNames('de_AT');
+        $names = Scripts::getNames('de_AT');
 
         foreach ($names as $script => $name) {
-            $this->assertSame($name, $this->dataProvider->getName($script));
+            $this->assertSame($name, Scripts::getName($script));
         }
     }
 }