+++ src/Symfony/Component/Validator/Constraints/BicValidator.php
@@ -11,7 +11,7 @@
 
 namespace Symfony\Component\Validator\Constraints;
 
-use Symfony\Component\Intl\Intl;
+use Symfony\Component\Intl\Regions;
 use Symfony\Component\PropertyAccess\Exception\NoSuchPropertyException;
 use Symfony\Component\PropertyAccess\PropertyAccess;
 use Symfony\Component\PropertyAccess\PropertyAccessor;
@@ -105,8 +105,8 @@ class BicValidator extends ConstraintValidator
         }
 
         // @deprecated since Symfony 4.2, will throw in 5.0
-        if (class_exists(Intl::class)) {
-            $validCountryCode = isset(Intl::getRegionBundle()->getCountryNames()[substr($canonicalize, 4, 2)]);
+        if (class_exists(Regions::class)) {
+            $validCountryCode = Regions::exists(substr($canonicalize, 4, 2));
         } else {
             $validCountryCode = ctype_alpha(substr($canonicalize, 4, 2));
             // throw new LogicException('The "symfony/intl" component is required to use the Bic constraint.');