+++ src/Symfony/Component/Intl/CHANGELOG.md
@@ -1,6 +1,15 @@
 CHANGELOG
 =========
 
+4.3.0
+-----
+
+ * deprecated `ResourceBundle` namespace
+ * added `Currencies` in favor of `Intl::getCurrencyBundle()`
+ * added `Languages` and `Scripts` in favor of `Intl::getLanguageBundle()`
+ * added `Locales` in favor of `Intl::getLocaleBundle()`
+ * added `Regions` in favor of `Intl::getRegionBundle()`
+
 4.2.0
 -----